requirements.txt
beautifulsoup4==4.6.0
certifi==2016.2.28
future==0.16.0
numpy==1.13.1
pandas==0.20.3
psutil==5.2.2
pymongo==3.4.0
python-dateutil==2.6.1
python-telegram-bot==8.1.1
pytz==2017.2
six==1.10.0
